chore(polish): Remove custom css validation
After some soul searching, decided to walk back custom CSS validation for this release. The highest
value items will be validation of the polished API itself, but not jumping down the rabbit hole of
trying to be a full on CSS validator. Especially since a lot of the code surrounding it won't be
validated.